backend/requirements.txt
flask
flask-cors
pymongo
google-generativeai
python-dotenv